+++ shuup/core/excs.py
@@ -56,6 +56,10 @@ class ProductNotVisibleProblem(Problem):
     pass
 
 
+class MissingSupplierModuleException(Exception):
+    pass
+
+
 class ImpossibleProductModeException(ValueError):
     def __init__(self, message, code=None):
         super(ImpossibleProductModeException, self).__init__(message)